Add logging entries and implement validation worker for topic checks
- Updated `document_service.log` and `garant_service.log` to include new log entries for initialization and server status.
- Introduced a new `validation_worker.go` file to handle validation requests for topics.
- Implemented logic to check the status of topics and log relevant information.
- Added error handling for validation requests and logging of results.

+++ garant_service/iternal/garant/check_modified.go
@@ -10,15 +10,15 @@ import (
 )
 
 type ModCheckRequest struct {
-	Topics     []int64 `json:"topics"`
-	ModDate    string  `json:"modDate"` // формат: "YYYY-MM-DD"
-	NeedEvents bool    `json:"needEvents"`
+	Topics     []int  `json:"topics"`
+	ModDate    string `json:"modDate"` // формат: "YYYY-MM-DD"
+	NeedEvents bool   `json:"needEvents"`
 }
 
 type ModCheckResponse struct {
 	Topics []struct {
-		Topic     int64 `json:"topic"`
-		ModStatus int   `json:"modStatus"` // 1 - изменился, 2 - не найден
+		Topic     int `json:"topic"`
+		ModStatus int `json:"modStatus"` // 1 - изменился, 2 - не найден
 		Events    []struct {
 			Date string `json:"date"`
 			Type int    `json:"type"` // тип события
@@ -26,7 +26,7 @@ type ModCheckResponse struct {
 	} `json:"topics"`
 }
 
-func CheckModified(topics []int64, sinceDate string) (*ModCheckResponse, error) {
+func CheckModified(topics []int, sinceDate string) (*ModCheckResponse, error) {
 	reqBody := ModCheckRequest{
 		Topics:     topics,
 		ModDate:    sinceDate,
